Add files via upload
Here is a possible description for a GitHub repository called "Stock Price Prediction And Forecasting Using Stacked LSTM- Deep Learning":

This repository contains a deep learning model for predicting stock prices using stacked LSTM networks. The model is trained on a dataset of historical stock prices and other financial data, and can be used to predict future stock prices for a variety of stocks.

The model is implemented using the TensorFlow library, and can be trained and used on both CPUs and GPUs. The repository also includes a Jupyter Notebook that demonstrates how to train and use the model.

This model can be used by investors and traders to make more informed investment decisions. It can also be used by researchers to study stock markets and develop new trading strategies.